… README.md
… UPGRADING.md
… jwt.test-d.ts
… package.json
… test/test.test.js