@radomird 2021-11-10
A new version of `fast-jwt` has been released. I've updated it and also addressed the issues mentioned in the PR. It's ready for reviews.

The following proposal from @darkgl0w :
> - We should take the opportunity offered by this PR to rework the file structure of the plugin and move jwt.test-d.ts to test/types/jwt.test-d.ts, rename test/test.test.js to test/jwt.test.js and update the corresponding references in package.json file.

falls out of scope of this PR so I'll create a separate ticket for that.