@simoneb 2021-11-10
> > LGTM with a little addition to the `README.md` note.
> > I spotted this breaking use case in the wild, but I guess this PR will land in a semver major. It's not a blocking case to me as I always use values expressed in milliseconds but this plugin is widely used so WDYT ? Is it worth adding this feature to `fast-jwt` before landing this PR ?
> 
> @mcollina @simoneb what do you think?

I'm not sure I understand what's the breaking use case mentioned by @darkgl0w