@darkgl0w 2021-11-10
> I'm not sure I understand what's the breaking use case mentioned by @darkgl0w

@simoneb An example use case that will break:

```javascript
  fastify.register(jwt, {
    secret: {
        private: 'private_key_here',
        public: 'public_key_here'
    },
    sign: {
      algorithm: 'RS256',
      // this will fail with this PR because string describing a time span are not converted in ms by `fast-jwt`
      expiresIn: '1 day'
    },
    verify: {
      algorithm: ['RS256']
    }
  });
```

After this PR it must be changed to something like this :
```javascript
  fastify.register(jwt, {
    secret: {
        private: 'private_key_here',
        public: 'public_key_here'
    },
    sign: {
      algorithm: 'RS256',
      expiresIn: 86_400_000 // 1 day in milliseconds
    },
    verify: {
      algorithm: ['RS256']
    }
  });
```