@jsumners 2021-11-10
> No strong opinion about whether to do it now or in an upcoming PR.

If this PR isn't a semver major, then this PR should include `ms` to handle the conversion so as not to make it a breaking change.