@darkgl0w 2021-11-10
> 
> 
> I think we should add support for `ms` in this library. It seems an unneeded breakage.

In this case it could be worth looking at [this fork of `ms`](https://github.com/lukeed/ms) in terms of performance. WDYT ?