@ShogunPanda 2021-11-11
@mcollina I've reviewed it. I haven't found anything else than what you already mentioned. I do agree that all signers and verifiers must be reused as much as we can. But given that options in fastify-jwt decorators can be changed at any call, this might be problematic. We might have to maintain a map of signers/verifiers/decoders according to the provided options.

About `ms` support, I do also agree that it should be added back. At the end of the day it's just intercepting those options before forwarding them to `fast-jwt`.