@mcollina 2021-11-11
> > I think we should add support for `ms` in this library. It seems an unneeded breakage.
> 
> In this case it could be worth looking at [this fork of `ms`](https://github.com/lukeed/ms) in terms of performance. WDYT ?

I don't think `ms` should be in the hot path for this module. If it is we have a problem.