@radomird 2021-11-15
Following the comment of @ShogunPanda :

> @mcollina I've reviewed it. I haven't found anything else than what you already mentioned. I do agree that all signers and verifiers must be reused as much as we can. But given that options in fastify-jwt decorators can be changed at any call, this might be problematic. We might have to maintain a map of signers/verifiers/decoders according to the provided options.

I would suggest to have a separate PR for optimising the solution. I would open a draft PR for that, where we could discuss how to solve the issues that were brought up in this comment. @mcollina what do you think?