@mcollina 2021-11-16
> I would suggest to have a separate PR for optimising the solution. I would open a separate issue for that, where we could discuss how to solve the issues that were brought up in this comment. @mcollina what do you think?

No. Moving from jsonwebtoken to fast-jwt is _mostly_ about performance. Doing the change without any performance benefits is breakage for our users without much gain.