@radomird 2021-11-16
Added conversion to milliseconds using the proposed library [@lukeed/ms](https://github.com/lukeed/ms) (proposed by @darkgl0w ). I've tried to minimise the impact on the hot path.