@radomird 2021-11-16
@mcollina I need your input before I push any changes regarding the optimisation. Should I go for:
- _simple solution_ - check if there are new options provided and then if yes, then create a new instance of signer/decoder/verifier and if not then we use the "global" one
OR
- _complicated solution_ - keep a global options cache and then cache all new options - in order to do this I might introduce a function for hashing the options object in order to keep track of it

I personally would prefer the simpler solution, as in all honesty, the complicated seems like an overkill.