@radomird 2021-11-16
> @radomird Aren't there existing memoization libraries that can handle per-parameter combination caching for you?

Indeed there are libraries for that (my preference [fast-memoize.js ](https://github.com/caiogondim/fast-memoize.js)) but I wanted Matteo's opinion of whether we should go that route, as then we include a new dependency to the library.