@mcollina 2021-11-17
The simple solution described above is ok.