@radomird 2021-11-17
Simple solution implemented