@radomird 2021-11-17
I've ran some manual benchmarks, comparing the old (using `jsonwebtoken`) implementation and the new (using `fast-jwt`). 

I did **only a limited number of runs** (7) and took the average of that and compared. So, **this is by no means a real thorough benchmark** (with 1000's of calls), but it gives an idea of the improvements.

#### OLD (`jsonwebtoken` implementation)
| method | 1 | 2 | 3 | 4 | 5 | 6 | 7 | AVG (not incl. the 1st run) |
| ------ | :---: | :---: | :---: | :---: | :---: | :---: | :---: | ---------------------------: |
| sign | 6.797 | 3.063 | 3.127 | 1.158 | 2.933 | 3.506 | 4.209 | **2.999 ms**
| jwtSign | 7.216 | 3.822 | 3.106 | 4.750 | 4.801 | 2.922 | 3.665 | **3.844 ms**
| verify | 3.077 | 0.347 | 0.380 | 0.380 | 0.403 | 0.536 | 0.424 | **0.411 ms**
| jwtVerify | 1.862 | 0.509 | 0.405 | 0.652 | 0.384 | 0.502 | 0.509 | **0.493 ms**

#### NEW (`fast-jwt` implementation)
| method | 1 | 2 | 3 | 4 | 5 | 6 | 7 | AVG (not incl. the 1st run) |
| ------ | :---: | :---: | :---: | :---: | :---: | :---: | :---: | ---------------------------: |
| sign | 6.499 | 2.130 | 2.106 | 3.147 | 2.145 | 2.039 | 2.162 | **2.288 ms**
| jwtSign | 8.605 | 2.024 | 2.025 | 2.861 | 2.058 | 2.064 | 3.147 | **2.363 ms**
| verify | 1.413 | 0.280 | 0.285 | 0.274 | 0.288 | 0.257 | 0.256 | **0.273 ms**
| jwtVerify | 1.533 | 0.415 | 0.288 | 0.386 | 0.306 | 0.286 | 0.259 | **0.323 ms**

#### Improvements (NEW vs OLD) based on the runs above
| method | improvement |
| ------ | -----------: |
| sign | 23% |
| jwtSign | 38% |
| verify | 33% |
| jwtVerify | 34% |